+++ src/src/main/java/io/codelex/classesandobjects/practice/BankAccount/BankBalance.java
@@ -1,27 +1,6 @@
-package io.codelex.classesandobjects.practice;
+package io.codelex.classesandobjects.practice.BankAccount;
 
 import java.util.Scanner;
-
-public class BankAccount {
-    public String name;
-    public double balance;
-
-    public final void deposit(double amount) {
-        balance += amount;
-    }
-
-    public final void withdraw(double amount) {
-        balance -= amount;
-    }
-
-    public String toString() {
-        if (balance >= 0) {
-            return name + ", $" + String.format("%.2f", balance);
-        } else {
-            return name + ", -$" + String.format("%.2f", -balance);
-        }
-    }
-}
 
 class BankBalance {
     public static void main(String[] args) {
