+++ src/src/main/java/io/codelex/classesandobjects/practice/Car/LitersPerKilometer.java
@@ -1,37 +1,6 @@
-package io.codelex.classesandobjects.practice;
+package io.codelex.classesandobjects.practice.Car;
 
 import java.util.Scanner;
-
-public class Car {
-    private double startKilometers;
-    private double endKilometers;
-    private double liters;
-
-    public Car(double startOdo, double endOdo, double liters) {
-        this.startKilometers = startOdo;
-        this.endKilometers = endOdo;
-        this.liters = liters;
-    }
-
-    public double calculateConsumption() {
-        return 100 * (liters / (endKilometers - startKilometers));
-    }
-
-    public boolean gasHog() {
-        double consumption = calculateConsumption();
-        return consumption > 15.0;
-    }
-
-    public boolean economyCar() {
-        double consumption = calculateConsumption();
-        return consumption < 5.0;
-    }
-
-    public void fillUp(int mileage, double liters) {
-        this.endKilometers = mileage;
-        this.liters += liters;
-    }
-}
 
 class LitersPerKilometer {
     public static void main(String[] args) {
@@ -61,4 +30,3 @@ class LitersPerKilometer {
         System.out.println("Car fill up: " + car.calculateConsumption());
     }
 }
-
