+++ src/src/main/java/io/codelex/classesandobjects/practice/Date/Date.java
@@ -1,0 +1,42 @@
+package io.codelex.classesandobjects.practice.Date;
+
+public class Date {
+    private int month;
+    private int day;
+    private int year;
+
+    public Date(int month, int day, int year) {
+        this.month = month;
+        this.day = day;
+        this.year = year;
+    }
+
+    public int getMonth() {
+        return month;
+    }
+
+    public void setMonth(int month) {
+        this.month = month;
+    }
+
+    public int getDay() {
+        return day;
+    }
+
+    public void setDay(int day) {
+        this.day = day;
+    }
+
+    public int getYear() {
+        return year;
+    }
+
+    public void setYear(int year) {
+        this.year = year;
+    }
+
+    public void displayDate() {
+        System.out.println(month + "/" + day + "/" + year);
+    }
+}
+
