+++ src/src/main/java/io/codelex/classesandobjects/practice/Movie/Movie.java
@@ -1,4 +1,4 @@
-package io.codelex.classesandobjects.practice;
+package io.codelex.classesandobjects.practice.Movie;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -40,18 +40,3 @@ public class Movie {
 
 }
 
-class PG_Movies {
-    public static void main(String[] args) {
-        Movie[] movies = {
-                new Movie("Casino Royale", "Eon Productions", "PG13"),
-                new Movie("Glass", "Buena Vista International", "PG13"),
-                new Movie("Spider-Man: Into the Spider-Verse", "Columbia Pictures", "PG")
-        };
-
-        Movie[] pgMovies = Movie.getPG(movies);
-        for (Movie movie : pgMovies) {
-            System.out.println(movie.getTitle());
-            System.out.println(movie.getStudio());
-        }
-    }
-}
