+++ src/src/main/java/io/codelex/classesandobjects/practice/Product/Product.java
@@ -1,4 +1,4 @@
-package io.codelex.classesandobjects.practice;
+package io.codelex.classesandobjects.practice.Product;
 
 public class Product {
     private final String name;
@@ -24,18 +24,4 @@ public class Product {
         this.amountAtStart = amountAtStart;
     }
 
-    public static void main(String[] args) {
-        Product firstProduct = new Product("Logitech mouse", 70.00, 14);
-        firstProduct.printProduct();
-
-        Product secondProduct = new Product("iPhone 5s", 999.99, 3);
-        secondProduct.printProduct();
-
-        Product thirdProduct = new Product("Epson EB-U05", 440.46, 1);
-        thirdProduct.printProduct();
-
-        thirdProduct.setAmountAtStart(99);
-        thirdProduct.setPriceAtStart(399.39);
-        thirdProduct.printProduct();
-    }
 }
