+++ src/src/main/java/io/codelex/classesandobjects/practice/SavingsAccount/BalanceTest.java
@@ -1,37 +1,6 @@
-package io.codelex.classesandobjects.practice;
+package io.codelex.classesandobjects.practice.SavingsAccount;
 
 import java.util.Scanner;
-
-public class SavingsAccount {
-    private double annualInterestRate;
-    private double balance;
-
-    public SavingsAccount(double initialBalance) {
-        balance = initialBalance;
-    }
-
-    public void Withdraw(double amount) {
-        balance -= amount;
-    }
-
-    public void Deposit(double amount) {
-        balance += amount;
-    }
-
-    public void addMonthlyInterest() {
-        double monthlyInterestRate = annualInterestRate / 12;
-        double monthlyInterest = balance * monthlyInterestRate;
-        balance += monthlyInterest;
-    }
-
-    public double getBalance() {
-        return balance;
-    }
-
-    public void setAnnualInterestRate(double rate) {
-        annualInterestRate = rate;
-    }
-}
 
 class BalanceTest {
     public static void main(String[] args) {
